requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
blas=1.0=mkl
ca-certificates=2021.7.5=haa95532_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py38haa95532_0
charset-normalizer=2.0.4=pypi_0
click=8.0.1=pypi_0
colorama=0.4.4=pypi_0
cudatoolkit=10.2.89=h74a9793_1
filelock=3.0.12=pypi_0
freetype=2.10.4=hd328e21_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.40.0=pypi_0
huggingface-hub=0.0.17=pypi_0
idna=3.2=pypi_0
intel-openmp=2021.3.0=haa95532_3372
joblib=1.0.1=pypi_0
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=pypi_0
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.2=py38hf11a4ad_0
ninja=1.10.2=h6d14046_1
numpy=1.20.3=py38ha4e8547_0
numpy-base=1.20.3=py38hc2deb75_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1l=h2bbff1b_0
packaging=21.0=pypi_0
pandas=1.3.3=pypi_0
pillow=8.3.1=py38h4fa10fc_0
pip=21.0.1=py38haa95532_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pypi_0
python=3.8.11=h6244533_1
python-dateutil=2.8.2=pypi_0
pytorch=1.9.0=py3.8_cuda10.2_cudnn7_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
regex=2021.8.28=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
sacremoses=0.0.45=pypi_0
setuptools=52.0.0=py38haa95532_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tk=8.6.10=he774522_0
tokenizers=0.10.3=pypi_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu102
tqdm=4.62.2=pypi_0
transformers=4.10.2=pypi_0
typing_extensions=3.10.0.0=pyhca03da5_0
urllib3=1.26.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
